requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
pydantic==2.5.2
aiohttp==3.9.1
python-multipart==0.0.6
motor==3.3.2
pymongo==4.6.1 